% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_aux_area.R
\name{lsm_aux_area}
\alias{lsm_aux_area}
\title{Calculate area}
\usage{
lsm_aux_area(
  input_null,
  input_id,
  area_round_digit = 0,
  area_unit = "ha",
  map_ncell = FALSE,
  table_export = FALSE
)
}
\arguments{
\item{input_null}{\verb{[character=""]} Habitat map (binary classification: e.g., 1/0 or 1/NA) in GRASS.}

\item{input_id}{\verb{[character=""]} Habitat map (binary classification: e.g., 1/0 or 1/NA) in GRASS.}

\item{area_round_digit}{\verb{[integer]} Decimal digits for area rounding.}

\item{area_unit}{\verb{[character=""]} Area unit: \code{"ha"}, \code{"m2"}, or \code{"km2"}.}

\item{map_ncell}{\verb{[logical]} Calculate number of cells.}

\item{table_export}{\verb{[logical]} Calculate number of cells.}
}
\description{
Calculate area.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "degrees")

# plot
plot(r, col = c("white", "forestgreen"), legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# null
rgrass::execGRASS(cmd = "r.mapcalc", flags = "overwrite", expression = "r_null = if(r == 1, 1, null())")

# clump
rgrass::execGRASS(cmd = "r.clump", flags = "overwrite", input = "r_null", output = "r_id")

# area
lsmetrics::lsm_aux_area(input_null = "r_null",
                        input_id = "r_id",
                        area_round_digit = 0,
                        area_unit = "ha",
                        map_ncell = TRUE,
                        table_export = TRUE)

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# plot
r_id <- terra::rast(rgrass::read_RAST("r_id", flags = "quiet", return_format = "SGDF"))

plot(r_id, legend = FALSE, axes = FALSE, main = "Fragment ID")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_id)

r_ncell <- terra::rast(rgrass::read_RAST("r_ncell", flags = "quiet", return_format = "SGDF"))

plot(r_ncell, legend = FALSE, axes = FALSE, main = "Number of cells")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_ncell)

r_area <- terra::rast(rgrass::read_RAST("r_area", flags = "quiet", return_format = "SGDF"))

plot(r_area, legend = FALSE, axes = FALSE, main = "Area (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_area, cex = .7, digits = 3)

# table
r_table_area <- vroom::vroom("r_table_area.csv", show_col_types = FALSE)
r_table_area

# delete table and grassdb
unlink("r_table_area.csv")
unlink("grassdb", recursive = TRUE)
}
