% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_aux_fill_hole.R
\name{lsm_aux_fill_hole}
\alias{lsm_aux_fill_hole}
\title{Fill holes}
\usage{
lsm_aux_fill_hole(
  input,
  output = NULL,
  zero_as_null = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification (e.g. values 1,0 or 1,NA for habitat,non-habitat) inside GRASS Data Base.}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_null}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero cells to null cells.}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}
}
\description{
Fill holes
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "meters")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# fill holes
lsmetrics::lsm_aux_fill_hole(input = "r")

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_aux_fill_hole <- rgrass::read_RAST("r_aux_fill_hole", flags = "quiet", return_format = "terra")

# plot
plot(r_aux_fill_hole, legend = FALSE, axes = FALSE, main = "Fragment fill hole")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_aux_fill_hole)

# delete grassdb
unlink("grassdb", recursive = TRUE)

}
