% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ lsm_buffer_statistics.R
\name{lsm_buffer_statistics}
\alias{lsm_buffer_statistics}
\title{Create grids and calculate statistics}
\arguments{
\item{input}{\verb{[character=""]} \cr}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{landscape_metric}{\verb{[character=""]} \cr}

\item{landscape_metric_has_null}{\verb{[character=""]} \cr}

\item{points}{\verb{[character=""]} \cr}

\item{distance}{\verb{[character=""]} \cr}

\item{column_prefix}{\verb{[character=""]} \cr}

\item{method}{\verb{[character=""]} \cr Univariate statistics: number, null_cells,
minimum ,maximum, range, average, stddev, variance, coeff_var, sum, first_quartile
,median, third_quartile, percentile}

\item{percentile}{\verb{[character=""]} \cr}
}
\description{
Create grids and calculate statistics
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- as.character(link2GI::findGRASS()[1]) # windows users need to find, e.g. "C:/Program Files/GRASS GIS 8.2"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# points
p <- terra::vect(data.frame(x = c(235150, 234450, 235150),
                            y = c(7525250, 7524850, 7524350)),
                 geom = c("x", "y"),
                 crs = "+proj=utm +zone=23 +south +datum=WGS84 +units=m +no_defs +type=crs")

# import raster from r to grass
rgrass::write_VECT(x = p, flags = c("o", "overwrite", "quiet"), vname = "p")

# buffer
lsmetrics::lsm_buffer_statistics(input = "r",
                                 landscape_metric = "r",
                                 landscape_metric_has_null = FALSE,
                                 point = "p",
                                 distance = 200,
                                 column_prefix = "percentage",
                                 method = "average")

# files
# rgrass::execGRASS(cmd = "g.list", type = "vector")

# import r
v_buffer <- rgrass::read_VECT("r_p_buffer200", flags = "quiet")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Percentage (\%)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)
plot(v_buffer, "percentage_average", alpha = .5, add = TRUE)
plot(p, col = "gray30", cex = 3, add = TRUE)
text(v_buffer, col = "white", labels = "percentage_average", cex = .7)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
