% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_core.R
\name{lsm_core}
\alias{lsm_core}
\title{Identify the core of landscape elements}
\usage{
lsm_core(
  input,
  output = NULL,
  zero_as_null = FALSE,
  edge_depth,
  edge_contraction = "maximum",
  area_round_digits = 0,
  area_unit = "ha",
  map_core_id = FALSE,
  map_core_ncell = FALSE,
  map_core_area = FALSE,
  map_core_number = FALSE,
  map_core_area_index = FALSE,
  map_core_id_original = FALSE,
  map_core_ncell_original = FALSE,
  map_core_area_original = FALSE,
  map_core_percentage = FALSE,
  buffer_radius = NULL,
  buffer_circular = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification (e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_null}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero cells to null cells.}

\item{edge_depth}{\verb{[numeric]} \cr Integer indicating edge distance in meters considered adjacent to form a patch.}

\item{edge_contraction}{\verb{[character=""]} \cr}

\item{area_round_digits}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{area_unit}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{map_core_id}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_ncell}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_area}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_area_index}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_id_original}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_ncell_original}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_area_original}{\verb{[logical(1)=FALSE]} \cr}

\item{map_core_percentage}{\verb{[logical(1)=FALSE]} \cr}

\item{buffer_radius}{\verb{[numeric]} \cr}

\item{buffer_circular}{\verb{[logical(1)=FALSE]} \cr}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}
}
\description{
Identify the core of landscape elements for each pixel. Calculate area, original area and percentage.
}
