% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_distance.R
\name{lsm_distance}
\alias{lsm_distance}
\title{Calculate patch distance}
\usage{
lsm_distance(input, output = NULL, zero_as_na = FALSE, type)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Patch area map name inside GRASS Data Base.}

\item{zero_as_na}{\verb{[logical=""]} \cr}

\item{type}{\verb{[character=""]} \cr}
}
\description{
Calculate focal ("moving window") values for each cell using the mean
from \link{r.neighbors} module and multiplies by 100.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- as.character(link2GI::findGRASS()[1])
path_grass

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite"), vname = "r")

# distance
lsm_distance(input = "r", zero_as_na = FALSE, type = "inside")
lsm_distance(input = "r", zero_as_na = FALSE, type = "outside")

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_dist_in <- rgrass::read_RAST("r_distance_inside", return_format = "terra")
r_dist_in

r_dist_out <- rgrass::read_RAST("r_distance_outside", return_format = "terra")
r_dist_out

# plot
plot(r_dist_in, legend = FALSE, axes = FALSE, main = "Distance inside (m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_dist_in)

plot(r_dist_out, legend = FALSE, axes = FALSE, main = "Distance outside (m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_dist_out)
}
