% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_distance_enn.R
\name{lsm_distance_enn}
\alias{lsm_distance_enn}
\title{Calculate fragment euclidean nearest neighbor distance}
\usage{
lsm_distance_enn(
  input,
  output = NULL,
  zero_as_null = FALSE,
  id_direction = 8,
  distance_round_digit = 0,
  grid_size = 10000,
  distance_radius = 2000,
  table_distance = FALSE
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Fragment fragment euclidean nearest neighbor distance map name inside GRASS Data Base.}

\item{zero_as_null}{\verb{[logical=""]} \cr}

\item{table_distance}{\verb{[logical=""]} \cr}

\item{id_directions}{\verb{[numerical=""]} \cr}
}
\description{
Calculate euclidean nearest neighbor distance among fragments in meters using
\link{r.clump} and \link{r.distance} GRASS GIS module.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "meters")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r")

# distance
lsmetrics::lsm_distance_enn(input = "r")

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_distance_enn <- terra::rast(rgrass::read_RAST("r_distance_enn", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_distance_enn, legend = FALSE, axes = FALSE, main = "Euclidean nearest neighbor distance (m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_distance_enn, cex = .5)

# delete grassdb
unlink("grassdb", recursive = TRUE)

}
