% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_diversity.R
\name{lsm_diversity}
\alias{lsm_diversity}
\title{Calculate landscape diversity}
\usage{
lsm_diversity(
  input,
  output = NULL,
  region_input = FALSE,
  buffer_radius,
  buffer_circular = FALSE,
  diversity_index,
  alpha = NULL,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Habitat map output name inside GRASS Data Base.}

\item{buffer_radius}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{diversity_index}{\verb{[character=""]} \cr Integer indicating window size.}

\item{alpha}{\verb{[character=""]} \cr Integer indicating window size.}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}

\item{buffer_cirular}{\verb{[logical=""]} \cr}
}
\description{
Calculate landscape diversity using \link{r.mapcalc} and \link{r.neighbors} modules.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "meters", values_type = "multiclass")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Classes")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite"), vname = "r")

# diversity
lsmetrics::lsm_diversity(input = "r", diversity_index = "simpson", buffer_radius = 100)

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_div_buf100 <- rgrass::read_RAST("r_diversity_shannon_buffer100", flags = c("c", "quiet"), return_format = "terra")

# plot
plot(r_div_buf100, legend = FALSE, axes = FALSE, main = "Landscape diversity (Shannon) (buffer 100 m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
text(r_div_buf100, digits = 1, cex = .75)

# delete grassdb
unlink("grassdb", recursive = TRUE)

}
