% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_diversity_parallel.R
\name{lsm_diversity_parallel}
\alias{lsm_diversity_parallel}
\title{Calculate landscape diversity}
\usage{
lsm_diversity_parallel(
  input,
  output = NULL,
  buffer_radius,
  index,
  alpha = NULL,
  grid_size,
  grid_delete = TRUE,
  install_rdiversity = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Habitat map output name inside GRASS Data Base.}

\item{buffer_radius}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{index}{\verb{[character=""]} \cr Integer indicating window size.}

\item{grid_size}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{grid_delete}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{install_rdiversity}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}
}
\description{
Calculate landscape diversity using \link{r.diversity} module.
}
