% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_enn_distance_fragment.R
\name{lsm_enn_distance_fragment}
\alias{lsm_enn_distance_fragment}
\title{Calculate fragment euclidean nearest neighbor distance}
\usage{
lsm_enn_distance_fragment(
  input,
  output = NULL,
  directions = 8,
  zero_as_na = FALSE,
  export_table = FALSE,
  buffer_fragment = 3000,
  nprocs = 1
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Fragment fragment euclidean nearest neighbor distance map name inside GRASS Data Base.}

\item{directions}{\verb{[numerical=""]} \cr}

\item{zero_as_na}{\verb{[logical=""]} \cr}

\item{export_table}{\verb{[logical=""]} \cr}

\item{buffer_fragment}{\verb{[numerical=""]} \cr}

\item{nprocs}{\verb{[numerical=""]} \cr}
}
\description{
Calculate euclidean nearest neighbor distance among fragments in meters using
\link{r.clump} and \link{r.distance} GRASS GIS module.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r")

# distance
lsmetrics::lsm_enn_distance_fragment(input = "r", zero_as_na = FALSE)

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_enn_distance <- terra::rast(rgrass::read_RAST("r_enn_distance", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_enn_distance, legend = FALSE, axes = FALSE, main = "Euclidean nearest neighbor distance (m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_enn_distance, cex = .5)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
