% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_functional_connectivity.R
\name{lsm_functional_connectivity}
\alias{lsm_functional_connectivity}
\title{Calculate functional connectivity}
\usage{
lsm_functional_connectivity(
  input,
  output = NULL,
  zero_as_na = FALSE,
  input_distance_outside,
  gap_crossing
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Patch area map name inside GRASS Data Base}

\item{gap_crossing}{\verb{[numeric(1)=4]\{4,8\}} \cr Integer indicating which cells are}

\item{resolution}{\verb{[numeric(1)=4]\{4,8\}} \cr Integer indicating which cells are}
}
\description{
Identifies functional patches connected.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- as.character(link2GI::findGRASS()[1])
path_grass

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite"), vname = "r")

# inside distance
lsm_distance(input = "r", zero_as_na = FALSE, type = "outside")

# functional connectivity
lsm_functional_connectivity(input = "r",
                            zero_as_na = FALSE,
                            input_distance_outside = "r_distance_outside",
                            gap_crossing = 100)

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import do r
r_confun100_pid <- rgrass::read_RAST("r_confun100_pid", return_format = "terra")
r_confun100_pid

plot(r_confun100_pid, legend = FALSE, axes = FALSE, main = "PID")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_confun100_pid)

# import habitat patch area to r
r_confun100_area <- rgrass::read_RAST("r_confun100_area_ha", return_format = "terra")
r_confun100_area

plot(r_confun100_area, legend = FALSE, axes = FALSE, main = "Area (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_confun100_area)

}
