% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_morphology.R
\name{lsm_landscape_morphology}
\alias{lsm_landscape_morphology}
\title{Identifies landscape morphologies}
\usage{
lsm_landscape_morphology(
  input,
  output = NULL,
  zero_as_null = FALSE,
  region_input = FALSE,
  morphological_segmentation = "all",
  table_morphological_segmentation = FALSE,
  area_round_digit = 0,
  area_unit = "ha",
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification  (e.g. values 1,0 or 1,NA for habitat,non-habitat) inside GRASS Data Base.}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{morphological_segmentation}{\verb{[character=""]} \cr}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}

\item{zero_as_na}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero cells to null cells.}
}
\description{
Identifies landscape morphologies: matrix, core, edge, corridor, stepping stone, branch, and perforation
}
