% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_patch_area.R
\name{lsm_patch_area}
\alias{lsm_patch_area}
\title{Calculate patch area}
\usage{
lsm_patch_area(
  input,
  output = NULL,
  zero_as_na = FALSE,
  id = FALSE,
  ncell = FALSE,
  area_integer = FALSE,
  patch_original = FALSE,
  patch_number = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat) inside GRASS Data Base.}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_na}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats
non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero
cells to null cells.}

\item{id}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{ncell}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{area_integer}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{patch_original}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{patch_number}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}
}
\description{
Identifies patch and calculates area in hectare.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# patch area
lsmetrics::lsm_patch_area(input = "r", id = TRUE, patch_original = TRUE, patch_number = TRUE)

# files
# rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_patch_id <- rgrass::read_RAST("r_patch_id", flags = "quiet", return_format = "terra")

plot(r_patch_id, legend = FALSE, axes = FALSE, main = "Patch id")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_patch_id)

r_patch_area <- rgrass::read_RAST("r_patch_area_ha", flags = "quiet", return_format = "terra")

plot(r_patch_area, legend = FALSE, axes = FALSE, main = "Patch area (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_patch_area)

r_patch_area_original <- rgrass::read_RAST("r_patch_area_ha_original", flags = "quiet", return_format = "terra")

plot(r_patch_area_original, legend = FALSE, axes = FALSE, main = "Patch area original (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_patch_area_original)

r_patch_number_original <- rgrass::read_RAST("r_patch_number_original", flags = "quiet", return_format = "terra")

plot(r_patch_number_original, legend = FALSE, axes = FALSE, main = "Number of patch (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_patch_number_original)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
