% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_percentage_parallel.R
\name{lsm_percentage_parallel}
\alias{lsm_percentage_parallel}
\title{Calculate percentage using parallel computing}
\usage{
lsm_percentage_parallel(
  input,
  output = NULL,
  zero_as_null = FALSE,
  buffer_radius,
  buffer_circular = FALSE,
  grid_size,
  grid_delete = TRUE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Habitat map output name inside GRASS Data Base.}

\item{zero_as_null}{\verb{[logical=""]} \cr}

\item{buffer_radius}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{grid_size}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}

\item{buffer_cirular}{\verb{[logical=""]} \cr}
}
\description{
Calculate focal ("moving window") values for each cell using the mean
from \link{r.neighbors} module and multiplies by 100.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "meters")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite"), vname = "r")

# percentage
lsmetrics::lsm_percentage_parallel(input = "r",
                                   buffer_radius = 100,
                                   grid_size = 1000,
                                   grid_delete = FALSE,
                                   nprocs = 1,
                                   memory = 300)

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")
rgrass::execGRASS(cmd = "g.list", type = "vector")

# import from grass to r
v <- rgrass::read_VECT("r_grid", flags = "quiet")
r_pct_buf100 <- rgrass::read_RAST("r_pct_buf100", flags = "quiet", return_format = "terra")

# plot
plot(r_pct_buf100, legend = FALSE, axes = FALSE, main = "Habitat percentage (buffer 100 m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
plot(v, lwd = 3, border = "blue", add = TRUE)
text(v, cex = 3, col = "blue")
text(r_pct_buf100, cex = .75)

# delete grassdb
unlink("grassdb", recursive = TRUE)

}
