% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_perimeter.R
\name{lsm_perimeter}
\alias{lsm_perimeter}
\title{Calculate perimeter}
\usage{
lsm_perimeter(input, output = NULL, zero_as_na = FALSE)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat) inside GRASS Data Base.}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_na}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats
non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero
cells to null cells.}
}
\description{
Calculate perimeter in meter.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- as.character(link2GI::findGRASS()[1]) # windows users need to find, e.g. "C:/Program Files/GRASS GIS 8.2"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# perimeter
lsmetrics::lsm_perimeter(input = "r", zero_as_na = TRUE)

# files
# rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_perimeter <- terra::rast(rgrass::read_RAST("r_perimeter", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_perimeter, legend = FALSE, axes = FALSE, main = "Perimeter (m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_perimeter, cex = .5)

# import from grass to r
r_perimeter_area_ratio <- terra::rast(rgrass::read_RAST("r_perimeter_area_ratio", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_perimeter_area_ratio, legend = FALSE, axes = FALSE, main = "Perimeter area ratio")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_perimeter_area_ratio, digits = 3, cex = .4)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
