% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_strutural_connectivity.R
\name{lsm_structural_connectivity}
\alias{lsm_structural_connectivity}
\title{Calculate structural connectivity}
\usage{
lsm_structural_connectivity(
  input,
  output = NULL,
  zero_as_na = FALSE,
  ncell = FALSE,
  area_integer = FALSE
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat) inside GRASS Data Base.}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_na}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats
non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero
cells to null cells.}

\item{ncell}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{area_integer}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{id}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}
}
\description{
Identifies patch and calculates area in hectare.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# structural connectivity
lsmetrics::lsm_structural_connectivity(input = "r")

# files
# rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_structural_connected_area <- rgrass::read_RAST("r_structural_connected_area", flags = "quiet", return_format = "terra")

plot(r_structural_connected_area, legend = FALSE, axes = FALSE, main = "Structural connected area (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_structural_connected_area)

# import from grass to r
r_structural_connectivity <- rgrass::read_RAST("r_structural_connectivity", flags = "quiet", return_format = "terra")

plot(r_structural_connectivity, legend = FALSE, axes = FALSE, main = "Structural connectivity (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_structural_connectivity)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
