% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_toy_landscape.R
\name{lsm_toy_landscape}
\alias{lsm_toy_landscape}
\title{Create a toy landscape}
\usage{
lsm_toy_landscape()
}
\description{
Create a toy landscape.
}
\examples{
library(lsmetrics)
library(terra)

# toy landscape
toy_landscape <- lsmetrics::lsm_toy_landscape()

# plot
plot(toy_landscape, legend = FALSE, axes = FALSE, main = "Toy landscape")
plot(as.polygons(toy_landscape, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(toy_landscape), add = TRUE)
text(toy_landscape)
}
