% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_edge.R
\name{lsm_edge}
\alias{lsm_edge}
\title{Identify edge of landscape elements}
\usage{
lsm_edge(
  input,
  output = NULL,
  zero_as_null = FALSE,
  edge_depth,
  edge_contraction = "maximum",
  area_round_digits = 0,
  area_unit = "ha",
  map_edge_id = FALSE,
  map_edge_ncell = FALSE,
  map_edge_area = FALSE,
  map_edge_id_original = FALSE,
  map_edge_ncell_original = FALSE,
  map_edge_area_original = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification (e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_null}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero cells to null cells.}

\item{edge_depth}{\verb{[numeric]} \cr Integer indicating edge distance in meters considered adjacent to form a patch.}

\item{edge_contraction}{\verb{[character=""]} \cr}

\item{area_round_digits}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{area_unit}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}}

\item{map_edge_id}{\verb{[logical(1)=FALSE]} \cr}

\item{map_edge_ncell}{\verb{[logical(1)=FALSE]} \cr}

\item{map_edge_area}{\verb{[logical(1)=FALSE]} \cr}

\item{map_edge_id_original}{\verb{[logical(1)=FALSE]} \cr}

\item{map_edge_ncell_original}{\verb{[logical(1)=FALSE]} \cr}

\item{map_edge_area_original}{\verb{[logical(1)=FALSE]} \cr}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}

\item{map_edge_percentage}{\verb{[logical(1)=FALSE]} \cr}

\item{buffer_radius}{\verb{[numeric]} \cr}

\item{buffer_circular}{\verb{[logical(1)=FALSE]} \cr}
}
\description{
Identify edge of landscape elements for each pixel. Calculate area, original area and percentage.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "meters")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r")

# core and edge
lsmetrics::lsm_core_edge(input = "r",
                         edge_depth = 100,
                         core_edge_type = "both",
                         id = TRUE,
                         core_number = TRUE,
                         core_edge_original = TRUE,
                         calculate_area = TRUE,
                         calculate_percentage = TRUE,
                         buffer_radius = 100,
                         buffer_circular = FALSE)

# files
# rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_core100 <- rgrass::read_RAST("r_core100", flags = "quiet", return_format = "terra")
r_core100_id <- rgrass::read_RAST("r_core100_id", flags = "quiet", return_format = "terra")
r_core100_area_ha <- rgrass::read_RAST("r_core100_area_ha", flags = "quiet", return_format = "terra")
r_core100_area_ha_original <- rgrass::read_RAST("r_core100_area_ha_original", flags = "quiet", return_format = "terra")
r_core100_core_number <- rgrass::read_RAST("r_core100_core_number_original", flags = "quiet", return_format = "terra")
r_core100_pct_buf100 <- rgrass::read_RAST("r_core100_pct_buf100", flags = "quiet", return_format = "terra")

r_edge100 <- rgrass::read_RAST("r_edge100", flags = "quiet", return_format = "terra")
r_edge100_id <- rgrass::read_RAST("r_edge100_id", flags = "quiet", return_format = "terra")
r_edge100_area_ha <- rgrass::read_RAST("r_edge100_area_ha", flags = "quiet", return_format = "terra")
r_edge100_area_ha_original <- rgrass::read_RAST("r_edge100_area_ha_original", flags = "quiet", return_format = "terra")
r_edge100_pct_buf100 <- rgrass::read_RAST("r_edge100_pct_buf100", flags = "quiet", return_format = "terra")

# plot
plot(r_core100, legend = FALSE, axes = FALSE, main = "Core")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_core100)

plot(r_core100_id, legend = FALSE, axes = FALSE, main = "Core id")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_core100_id)

plot(r_core100_area_ha, legend = FALSE, axes = FALSE, main = "Core area (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_core100_area_ha)

plot(r_core100_area_ha_original, legend = FALSE, axes = FALSE, main = "Core area original (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_core100_area_ha_original)

plot(r_core100_core_number, legend = FALSE, axes = FALSE, main = "Number of cores")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_core100_core_number)

plot(r_core100_pct_buf100, legend = FALSE, axes = FALSE, main = "Core percentage (buffer 100 m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_core100_pct_buf100, cex = .75)


plot(r_edge100, legend = FALSE, axes = FALSE, main = "Edge")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_edge100)

plot(r_edge100_id, legend = FALSE, axes = FALSE, main = "Edge id")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_edge100_id)

plot(r_edge100_area_ha, legend = FALSE, axes = FALSE, main = "Edge area (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_edge100_area_ha)

plot(r_edge100_area_ha_original, legend = FALSE, axes = FALSE, main = "Edge area original (ha)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_edge100_area_ha_original)

plot(r_edge100_pct_buf100, legend = FALSE, axes = FALSE, main = "Edge percentage (buffer 100 m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_edge100_pct_buf100, cex = .75)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
