% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_percentage.R
\name{lsm_percentage}
\alias{lsm_percentage}
\title{Calculate fragment percentage}
\usage{
lsm_percentage(
  input,
  output = NULL,
  zero_as_na = FALSE,
  buffer_radius,
  buffer_circular = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification
(e.g. values 1,0 or 1,NA for habitat,non-habitat).}

\item{output}{\verb{[character=""]} \cr Habitat map output name inside GRASS Data Base.}

\item{zero_as_na}{\verb{[logical=""]} \cr}

\item{buffer_radius}{\verb{[numeric()]} \cr Integer indicating window size.}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}

\item{buffer_cirular}{\verb{[logical=""]} \cr}
}
\description{
Calculate focal ("moving window") values for each cell using the mean
from \link{r.neighbors} module and multiplies by 100.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
f <- system.file("raster/toy_landscape_habitat.tif", package = "lsmetrics")
r <- terra::rast(f)

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- as.character(link2GI::findGRASS()[1]) # windows users need to find, e.g. "C:/Program Files/GRASS GIS 8.2"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite"), vname = "r")

# percentage
lsmetrics::lsm_percentage(input = "r", output = "1", buffer_radius = 100)

# files
# rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_pct_buf100 <- rgrass::read_RAST("r_pct_buf100", return_format = "terra")

# plot
plot(r_pct_buf100, legend = FALSE, axes = FALSE, main = "Habitat percentage (buffer 100 m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_pct_buf100, cex = .75)

# delete grassdb
unlink("grassdb", recursive = TRUE)
}
