% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_perimeter.R
\name{lsm_perimeter}
\alias{lsm_perimeter}
\title{Calculate perimeter}
\usage{
lsm_perimeter(
  input,
  output = NULL,
  zero_as_null = FALSE,
  region_input = FALSE,
  perimeter_round_digit = 0,
  map_perimeter_area_ratio_index = FALSE,
  map_shape_index = FALSE,
  map_fractal_index = FALSE,
  nprocs = 1,
  memory = 300
)
}
\arguments{
\item{input}{\verb{[character=""]} \cr Habitat map, following a binary classification (e.g. values 1,0 or 1,NA for habitat,non-habitat) inside GRASS Data Base.}

\item{output}{\verb{[character=""]} \cr Map name output inside GRASS Data Base.}

\item{zero_as_null}{\verb{[logical(1)=FALSE]} \cr If \code{TRUE}, the function treats non-habitat cells as null; if \code{FALSE}, the function converts non-habitat zero cells to null cells.}

\item{map_perimeter_area_ratio_index}{\verb{[character=""]} \cr If \code{TRUE}, the function treats.}

\item{map_shape_index}{\verb{[character=""]} \cr If \code{TRUE}, the function treats.}

\item{map_fractal_index}{\verb{[character=""]} \cr If \code{TRUE}, the function treats.#' @param nprocs \verb{[numeric()]} \cr}

\item{nprocs}{\verb{[numeric()]} \cr}

\item{memory}{\verb{[numeric()]} \cr}
}
\description{
Calculate perimeter in meters and shape index.
}
\examples{
library(lsmetrics)
library(terra)

# read habitat data
r <- lsmetrics::lsm_toy_landscape(proj_type = "meters")

# plot
plot(r, legend = FALSE, axes = FALSE, main = "Binary habitat")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r)

# find grass
path_grass <- system("grass --config path", inter = TRUE) # windows users need to find the grass gis path installation, e.g. "C:/Program Files/GRASS GIS 8.3"

# create grassdb
rgrass::initGRASS(gisBase = path_grass,
                  SG = r,
                  gisDbase = "grassdb",
                  location = "newLocation",
                  mapset = "PERMANENT",
                  override = TRUE)

# import raster from r to grass
rgrass::write_RAST(x = r, flags = c("o", "overwrite", "quiet"), vname = "r", verbose = FALSE)

# perimeter
lsmetrics::lsm_perimeter(input = "r",
                         map_perimeter_area_ratio_index = TRUE,
                         map_shape_index = TRUE,
                         map_fractal_index = TRUE)

# files
rgrass::execGRASS(cmd = "g.list", type = "raster")

# import from grass to r
r_perimeter <- terra::rast(rgrass::read_RAST("r_perimeter", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_perimeter, legend = FALSE, axes = FALSE, main = "Perimeter (m)")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_perimeter, cex = .5)

# import from grass to r
r_perimeter_area_ratio_index <- terra::rast(rgrass::read_RAST("r_perimeter_area_ratio_index", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_perimeter_area_ratio_index, legend = FALSE, axes = FALSE, main = "Perimeter-area ratio index")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_perimeter_area_ratio_index, digits = 3, cex = .4)

# import from grass to r
r_shape_index <- terra::rast(rgrass::read_RAST("r_shape_index", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_shape_index, legend = FALSE, axes = FALSE, main = "Shape index")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_shape_index, digits = 3, cex = .4)

# import from grass to r
r_fractal_index <- terra::rast(rgrass::read_RAST("r_fractal_index", flags = "quiet", return_format = "SGDF"))

# plot
plot(r_fractal_index, legend = FALSE, axes = FALSE, main = "Fractal dimension index")
plot(as.polygons(r, dissolve = FALSE), lwd = .1, add = TRUE)
plot(as.polygons(r), add = TRUE)
text(r_fractal_index, digits = 3, cex = .4)

# delete grassdb
unlink("grassdb", recursive = TRUE)

}
